mental-wellbeing-ai/backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
scikit-learn
pandas
numpy
python-dotenv
